+++ ext/storage/messages.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"time"
 
+	sous "github.com/opentable/sous/lib"
 	"github.com/opentable/sous/util/logging"
 )
 
@@ -32,7 +33,10 @@ func reportSQLMessage(log logging.LogSink, started time.Time, sql string, rowcou
 
 // DefaultLevel implements LogMessage on sqlMessage
 func (msg *sqlMessage) DefaultLevel() logging.Level {
-	return logging.InformationLevel
+	if msg.err == nil {
+		return logging.InformationLevel
+	}
+	return logging.WarningLevel
 }
 
 // Message implements LogMessage on sqlMessage
@@ -52,5 +56,81 @@ func (msg *sqlMessage) EachField(fn logging.FieldReportFn) {
 	fn("sous-sql-rows", msg.rowcount)
 	if msg.err != nil {
 		fn("sous-sql-errreturned", msg.err.Error())
+	}
+}
+
+type (
+	storeMessage struct {
+		logging.CallerInfo
+		logging.MessageInterval
+		direction direction
+		state     *sous.State
+		err       error
+	}
+
+	direction uint
+)
+
+const (
+	read direction = iota
+	write
+)
+
+func reportReading(log logging.LogSink, started time.Time, state *sous.State, err error) {
+	msg := newStoreMessage(started, read, state, err)
+	msg.CallerInfo.ExcludeMe()
+	logging.Deliver(msg, log)
+}
+
+func reportWriting(log logging.LogSink, started time.Time, state *sous.State, err error) {
+	msg := newStoreMessage(started, write, state, err)
+	msg.CallerInfo.ExcludeMe()
+	logging.Deliver(msg, log)
+}
+
+func newStoreMessage(started time.Time, dir direction, state *sous.State, err error) *storeMessage {
+	return &storeMessage{
+		CallerInfo:      logging.GetCallerInfo(logging.NotHere()),
+		MessageInterval: logging.NewInterval(started, time.Now()),
+		state:           state,
+		direction:       dir,
+		err:             err,
+	}
+}
+
+func (msg *storeMessage) DefaultLevel() logging.Level {
+	if msg.err == nil {
+		return logging.DebugLevel
+	}
+	return logging.WarningLevel
+}
+
+func (msg *storeMessage) Message() string {
+	return msg.direction.message()
+}
+
+func (dir direction) message() string {
+	switch dir {
+	default:
+		return "Unknown state storage direction (shouldn't ever occur?)"
+	case read:
+		return "Reading state"
+	case write:
+		return "Writing state"
+	}
+}
+
+func (msg *storeMessage) EachField(fn logging.FieldReportFn) {
+	fn("@loglov3-otl", "sous-storage")
+	msg.CallerInfo.EachField(fn)
+	msg.MessageInterval.EachField(fn)
+	if msg.err != nil {
+		fn("sous-storage-error", msg.err.Error())
+	}
+	deps, err := msg.state.Deployments()
+	if err == nil {
+		fn("sous-storage-deployments", deps.Len())
+	} else {
+		fn("sous-storage-deployments", 0)
 	}
 }