+++ ext/storage/postgresstatemanager_test.go
@@ -34,42 +34,18 @@ func SetupTest(t *testing.T) *PostgresStateManagerSuite {
 		require:    require.New(t),
 	}
 
+	db := setupDB(t)
+
+	sink, ctrl := logging.NewLogSinkSpy()
+	suite.manager = NewPostgresStateManager(db, sink)
+
+	suite.logs = ctrl
+
 	port := "6543"
 	if np, set := os.LookupEnv("PGPORT"); set {
 		port = np
 	}
-	connstr := fmt.Sprintf("dbname=sous_test_template host=localhost port=%s sslmode=disable", port)
-	setupDB, err := sql.Open("postgres", connstr)
-	if err != nil {
-		suite.FailNow("Error setting up test database", "Error: %v. Did you already `make postgres-test-prepare`?", err)
-	}
-	// ignoring error because I think "no such DB is a failure"
-	setupDB.Exec("drop database sous_test")
-	if _, err := setupDB.Exec("create database sous_test template sous_test_template"); err != nil {
-		suite.FailNow("Error creating test database", "connstr %q err %v", connstr, err)
-	}
-	if err := setupDB.Close(); err != nil {
-		suite.FailNow("Error closing DB manipulation connection", "connstr %q err %v", connstr, err)
-	}
-
-	sink, ctrl := logging.NewLogSinkSpy()
-
-	suite.manager, err = NewPostgresStateManager(PostgresConfig{
-		DBName:   "sous_test",
-		User:     "",
-		Password: "",
-		Host:     "localhost",
-		Port:     port,
-		SSL:      false,
-	}, sink)
-
-	suite.logs = ctrl
-
-	if err != nil {
-		suite.FailNow("Setting up", "error: %v", err)
-	}
-
-	connstr = fmt.Sprintf("dbname=sous_test host=localhost port=%s sslmode=disable", port)
+	connstr := fmt.Sprintf("dbname=sous_test host=localhost port=%s sslmode=disable", port)
 	if suite.db, err = sql.Open("postgres", connstr); err != nil {
 		suite.FailNow("Error establishing test-assertion DB connection.", "Error: %v", err)
 	}
@@ -84,7 +60,7 @@ func TestPostgresStateManagerWriteState_success(t *testing.T) {
 	suite.require.NoError(suite.manager.WriteState(s, testUser))
 	suite.Equal(int64(2), suite.pluckSQL("select count(*) from deployments"))
 
-	assert.Len(t, suite.logs.CallsTo("LogMessage"), 12)
+	assert.Len(t, suite.logs.CallsTo("LogMessage"), 13)
 	message := suite.logs.CallsTo("LogMessage")[0].PassedArgs().Get(1).(logging.LogMessage)
 	logging.AssertMessageFields(t, message, append(
 		append(logging.StandardVariableFields, logging.IntervalVariableFields...), "sous-sql-query", "sous-sql-rows"),