+++ CHANGELOG.md
@@ -10,6 +10,7 @@ with respect to its command line interface and HTTP interface.
 
 ### Added
 * CLI: If no image is present in runspec, return a fatal flaw in build.
+* Server: adding duplex state storage, to keep DB in sync until ready to switch over
 * Server: Update logging to a more structured format, singularity package also
 
 ### Changed